data_pipeline/requirements.txt
apache-airflow==2.9.3
pyspark==3.5.1
pandas>=2.0.0
pyyaml>=6.0
